rename: vertical 폴더에 공백제거
