+++ shared/ui/table/vertical/index.tsx
