+++ shared/ui/table/vertical/vertical.module.scss
